usage, command move object helpers to lang
Reorganize helper libs into logical modules

Semver: patch
Ref: v7.0.0
