+++ lib/command.js
@@ -17,28 +17,28 @@
  * @requires mout/object/hasOwn
  * @requires mout/lang/isObject
  * @requires seeli/lib/domain
- * @requires seeli/lib/helpers/commands
+ * @requires seeli/lib/usage
  * @requires seeli/lib/exceptions
  **/
 const  os           = require( 'os' )                 // native os
      , tty          = require( 'tty' )
      , url          = require( 'url' )
      , path         = require( 'path' )
+     , events       = require( 'events' )             // native events
+     , clone        = require( 'clone' )
      , inquirer     = require( 'inquirer' )           // native inquirer
      , nopt         = require( 'nopt' )               // native nopt
      , chalk        = require( 'chalk' )              // ansi color support
      , strip        = require( 'strip-ansi')          // function helper to remove ansi color
-     , events       = require( 'events' )             // native events
      , array_append = require( 'mout/array/append' )  //
      , toArray      = require( 'mout/lang/toArray' )  //
-     , domain       = require( './domain')
-     , debug        = require( 'debug')('seeli:command')
-     , helpers      = require( './helpers/commands' ) //
-     , object       = require( './helpers/object' )   // object helpers
-     , exceptions   = require( './exceptions')
      , hasOwn       = require( 'mout/object/hasOwn' )
-     , clone        = require( 'clone' )
      , isObject     = require( 'mout/lang/isObject' )
+     , debug        = require( 'debug')('seeli:command')
+     , domain       = require( './domain')
+     , usage        = require('./usage')
+     , object       = require( './lang/object' )
+     , exceptions   = require( './exceptions')
      , stop_flags   = ['help', 'interactive', 'skip','color']
      , ARGV         = 'argv'
      , on_exp       = /^on[A-z]/
@@ -190,7 +190,7 @@ class Command extends events.EventEmitter {
    * @type string
    **/
   get usage(){
-    var out = helpers.usage.from( this.options.flags );
+    var out = usage.from( this.options.flags );
     return array_append(
       toArray( this.options.usage ),
       [
