+++ lib/commands/help.js
@@ -6,11 +6,11 @@
  * @author Eric Satterwhite
  * @requires os
  * @requires seeli/lib/command
- * @requires seeli/lib/helpers/command
+ * @requires seeli/lib/usage/list
  **/
 var os      = require('os')
   , Command = require('../command')
-  , list = require('../helpers/commands').list
+  , list = require('../usage/list')
   ;
 
 /**
