--- lib/helpers/object/path-exp.js
@@ -1,2 +1,0 @@
-'use strict';
-module.exports = /^(.+)\:(.+)$/;
