+++ lib/lang/object/index.js
@@ -4,18 +4,17 @@
 /**
  * Object helpers
  * @author Eric Satterwhite
- * @since 6.0.0
- * @module seeli/lib/helpers/object
- * @requires seeli/lib/helpers/object/set
- * @requires seeli/lib/helpers/object/key-for
+ * @since 7.0.0
+ * @module seeli/lib/lang/object
+ * @requires seeli/lib/lang/object/set
  **/
 
 module.exports = {
   /**
    * @property set {Function} sets a deep property on an object
    * @type Function
-   * @memberof module:seeli/lib/helpers/object
-   * @see module:seeli/lib/helpers/object/set
+   * @memberof module:seeli/lib/lang/object
+   * @see module:seeli/lib/lang/object/set
    **/
   set: require('./set')
 };
