+++ lib/lang/object/path-exp.js
@@ -1,0 +1,9 @@
+'use strict';
+/**
+ * Regex for parsing nested object flags
+ * @module seel/lib/lang/object/path-exp
+ * @since v7.0.0
+ * @author Eric Satterwhie
+ * @exports RegExp
+ **/
+module.exports = /^(.+)\:(.+)$/;
