+++ lib/lang/object/set.js
@@ -3,9 +3,9 @@
 
 /**
  * Helper function settign nested properties on an object
- * @module seeli/lib/helpers/object/set
+ * @module seeli/lib/lang/object/set
  * @author Eric Satterwhite
- * @requires seeli/lib/helpers/object/path-exp
+ * @requires seeli/lib/lang/object/path-exp
  * @type Function
  * @param {Object} source The source object to set keys on
  * @param {String} key a key path using a colon to separate key levels
